Feature/session add basic logic

See merge request bkrawcz1/paint-23l!3

+++ SESSION/src/main/java/org/example/session/entity/GuessEntity.java
@@ -17,7 +17,7 @@ public class GuessEntity {
 	Long id;
 
 	@Column("SESSION_ID")
-	Integer sessionId;
+	Long sessionId;
 
 	@Column("GUESS_NUMBER")
 	Integer guessNumber;
